controller改为接收前台传递json的形式

+++ src/main/java/com/gs/reusebook/dao/GoodsDao.java
@@ -9,5 +9,5 @@ import com.gs.reusebook.bean.Goods;
 
 public interface GoodsDao {
 	public List<Goods> selectAndPagedByName
-	(@Param("name") String name, @Param("offset") long offset, @Param("limit") long limit);
+	(@Param("name") String name, @Param("offset") int offset, @Param("limit") int limit);
 }
